fix: make man page install optional (doc/ was removed)

+++ packages/proot/build.sh
@@ -23,8 +23,10 @@ termux_step_pre_configure() {
 }
 
 termux_step_post_make_install() {
-	mkdir -p $TERMUX_PREFIX/share/man/man1
-	install -m600 $TERMUX_PKG_SRCDIR/doc/proot/man.1 $TERMUX_PREFIX/share/man/man1/proot.1
+	if [ -f $TERMUX_PKG_SRCDIR/doc/proot/man.1 ]; then
+		mkdir -p $TERMUX_PREFIX/share/man/man1
+		install -m600 $TERMUX_PKG_SRCDIR/doc/proot/man.1 $TERMUX_PREFIX/share/man/man1/proot.1
+	fi
 
 	sed -e "s|@TERMUX_PREFIX@|$TERMUX_PREFIX|g" \
 		$TERMUX_PKG_BUILDER_DIR/termux-chroot \
